RMC/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asn1crypto=1.3.0=py27_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py27_0
cffi=1.13.2=py27hb5b8e2f_0
cryptography=2.8=py27ha12b0ac_0
enum34=1.1.6=py27_1
idna=2.8=py27_0
ipaddress=1.0.23=py_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
ncurses=6.1=h0a44026_1
openssl=1.1.1d=h1de35cc_3
pip=19.3.1=py27_0
pycparser=2.19=py_0
pymysql=0.9.3=py27_0
python=2.7.5=3
readline=6.2=2
setuptools=44.0.0=py27_0
six=1.13.0=py27_0
sqlite=3.30.1=ha441bb4_0
tk=8.5.18=0
wheel=0.33.6=py27_0
zlib=1.2.11=h1de35cc_3
